A Collection of Tools for VARMA and State Space Processes

This package provides models for stationary processes with a rational spectral density and methods for their estimation.

Author(s)

Wolfgang Scherrer and Bernd Funovits

Maintainer: bernd.funovits@gmail.com

See also

Author

Maintainer: Bernd Funovits bernd.funovits@gmail.com (ORCID)

Authors:

Wolfgang Scherrer wolfgang.scherrer@tuwien.ac.at