Outputs of a statespace system

This internal helper function computes the outputs and states for a statespace system of the form $$a_{t+1} = A a_t + B u_t, \; y_t = C a_t + D u_t$$

Arguments

A

\((s,s)\) matrix.

B

\((s,n)\) matrix.

C

\((m,s)\) matrix.

D

\((m,n)\) matrix.

u

\((n,N)\) matrix with the inputs/disturbances: \((u_1,u_2,\ldots,u_N)\).

a

\((s,N+1)\) matrix. This matrix is overwritten with the (computed) states: \((a_1,a_2,\ldots,a_N,a_{N+1})\). On input a[,1] must hold the initial state \(a_1\).

y

\((m,N)\) matrix. This matrix is overwritten with (computed) outputs: \((y_1,y_2,\ldots,y_N)\).

Value

This RcppArmadillo routine returns NULL but overwrites the input arguments a and u!

Note

Use this procedure with care!

The procedure does not check the input arguments.

The procedure overwrites the input arguments a and u.

The data matrices are organized columnwise (to avoid memory shuffling)!

See also

Examples

# generate a random statespace model (3 outputs, 2 inputs and 4 states)
m = 3
n = 2
s = 4
model = test_stspmod(dim = c(m, n), s = s, digits = 2)

# generate random noise sequence (sample size N = 10)
n.obs = 10
u = matrix(rnorm(n.obs*n), nrow = n, ncol = n.obs)
print(u)
#>              [,1]      [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] -0.008161641 0.4151075 -2.5076274 -1.1354969 -0.6171516 -0.6986756
#> [2,]  0.675443061 0.7733806 -0.9724242  0.2943615 -0.2111856  0.7124553
#>             [,7]       [,8]       [,9]       [,10]
#> [1,] -0.07804087  1.0509301  0.7876608  0.02725284
#> [2,]  1.09116552 -0.1836981 -0.4505679 -0.39597250

# generate matrix for the state sequence
a = matrix(0, nrow = s, ncol = n.obs+1)
a[,1] = rnorm(s) # random initial state a[1]
print(a)
#>             [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#> [1,]  0.82440488    0    0    0    0    0    0    0    0     0     0
#> [2,]  2.46481445    0    0    0    0    0    0    0    0     0     0
#> [3,] -0.09118371    0    0    0    0    0    0    0    0     0     0
#> [4,]  0.37613488    0    0    0    0    0    0    0    0     0     0

# generate matrix for the outputs
y = matrix(0, nrow = m, ncol = n.obs)

# call outputs_STSP_cpp()
outputs_STSP_cpp(model$sys$A, model$sys$B, model$sys$C, model$sys$D, u, a, y)
print(u)
#>              [,1]      [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] -0.008161641 0.4151075 -2.5076274 -1.1354969 -0.6171516 -0.6986756
#> [2,]  0.675443061 0.7733806 -0.9724242  0.2943615 -0.2111856  0.7124553
#>             [,7]       [,8]       [,9]       [,10]
#> [1,] -0.07804087  1.0509301  0.7876608  0.02725284
#> [2,]  1.09116552 -0.1836981 -0.4505679 -0.39597250
print(a)  # a is overwritten with the computed states
#>             [,1]      [,2]      [,3]       [,4]      [,5]       [,6]
#> [1,]  0.82440488 -2.293077  0.318788  -4.357114   2.67933  -16.13106
#> [2,]  2.46481445  3.987800  9.394199  16.745636  35.95045   83.80753
#> [3,] -0.09118371 -4.031825  4.759263  -9.383869  34.57869  -16.25677
#> [4,]  0.37613488 -2.471480 -6.150455 -25.503015 -39.91524 -122.74637
#>             [,7]        [,8]        [,9]      [,10]      [,11]
#> [1,]    3.704392  -66.068773   -17.05783  -280.9372  -215.8645
#> [2,]  189.927389  441.826741  1008.66844  2327.1169  5312.8353
#> [3,]  132.125047   -2.799223   531.20666   245.7235  2287.6675
#> [4,] -221.973413 -608.763961 -1211.59866 -3099.6949 -6528.2644
print(y)  # y is overwritten with the computed outputs
#>            [,1]       [,2]      [,3]      [,4]       [,5]      [,6]       [,7]
#> [1,] -2.3197032 -4.3081398 -6.460507 -21.35190  -6.501508 -80.62249  -65.86759
#> [2,]  4.9021958  2.8863811 36.229599  47.38768 200.519845 296.89768  966.01247
#> [3,] -0.4349683  0.8935802 -8.737062  10.33498 -44.407720  14.70241 -173.58398
#>            [,8]      [,9]     [,10]
#> [1,] -361.30857 -463.7478 -1730.969
#> [2,] 1690.28234 4795.0816  9329.061
#> [3,]  -31.58329 -721.1775  -489.352