Toeplitz Calculations

Multiplication of stacked data vector with a block Toeplitz matrix (toepl_fwd() for MA calculations) or "inversion" of a block Toeplitz matrix in order to perform calculations equivalent to multiplying a given stacked data vector with the inverse of a lower-triangular banded block Toeplitz matrix (toepl_inv() for AR calculations). Note that matrix polynomials can be mapped one-to-one to banded lower-triangular block Toeplitz matrices.

Usage

toepl_fwd(polm_wide, data_in, t0 = 1)

toepl_inv(polm_rev, data_in, t0 = 1)

Arguments

polm_wide

Wide matrix \(( d_0, d_1, \ldots, d_q )\) of dimension \((m \times (q+1) n)\) which represents a matrix polynomial \(d(z)\) of degree \(q\).

data_in

Data matrix of dimension \((dim(inputs) x nobs)\). Every column corresponds to one observation.

t0

Integer. Time index where calculations should start. Default set to 1. For AR calculations, \(degree + 1\) would be another smart option.

polm_rev

Wide matrix \((c_p, ... , c_1)\) of dimension \((n \times (q+1) n)\) with coefficients ordered in reverse direction, and zero-lag coefficient matrix not included. It represents a square polynomial matrix \(c(z)\) with \(c_0\) equal to the identity matrix and of degree \(p\).

Value

data_out Data matrix of dimension \((dim_out x n_obs)\)

MA-type Toeplitz calculations

Given a polynomial matrix of degree \(q\) and dimension \((m \times n)\), where \(m \geq n\), and given a "wide" input data matrix of dimension \((n \times nobs)\), where \(nobs\) is the number of observations such that each column corresponds to one observation and the number of columns is equal to the number of observations, we calculate a "wide" output data matrix of dimension \((m \times nobs)\).

The function name toepl_fwd stems from the multiplication of the "stacked" input data vector \((u_1', \ldots , u_{nobs}')'\) with a banded lower-triangular block Toeplitz matrix \(T\) of dimension \((nobs m \times nobs n)\) whose block elements depend only on the difference between the row- and column-index such that \(T_{i,j} = d_{i-j}\).

AR-type Toeplitz calcuations

Given a square polynomial matrix \(c(z)\) for which \(c_0\) is the identity matrix and given a wide data matrix y = data_in, obtain the solution u, a wide data matrix, of the Toeplitz equation $$T (y_1', \ldots , y_{nobs}')' = (u_1', ... , u_{nobs}')'$$

Note that the zero-lag coefficient is discarded and the coefficients are in reversed order since this simplifies computations and implementation.

Examples

p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_fwd(p, data)
#>          [,1]     [,2]      [,3]      [,4]     [,5]     [,6]       [,7]
#> [1,] 288.5814 163.5193 -129.2345 -204.1137 41.90193 259.8999  -49.80246
#> [2,] 533.4095 299.5213 -250.9234 -385.3586 56.28360 468.3999 -105.15109
#>           [,8]      [,9]    [,10]    [,11]     [,12]      [,13]      [,14]
#> [1,] -150.4690 -32.77850 186.7236 100.0098 -170.2282  -659.1878  -657.8683
#> [2,] -287.6102 -75.10211 330.9000 186.6325 -320.6437 -1224.0077 -1227.6038
#>          [,15]     [,16]     [,17]    [,18]     [,19]    [,20]    [,21]
#> [1,] -288.9650 -239.9395 -50.14275 164.5694  591.2922 478.5993 142.4752
#> [2,] -527.6538 -444.2707 -65.17002 315.1147 1099.8144 873.0329 252.2172
#>         [,22]     [,23]     [,24]    [,25]    [,26]    [,27]     [,28]    [,29]
#> [1,] 139.4921  72.66673 -112.8571 24.16490 148.3693 398.0647  95.88839 224.8765
#> [2,] 256.0160 127.83231 -200.4524 62.46022 291.1953 741.7842 178.77359 426.5986
#>         [,30]     [,31]     [,32]    [,33]    [,34]    [,35]   [,36]    [,37]
#> [1,] 22.70158 -128.5303 -323.6586 127.8545 325.5101 443.1030 201.518 271.4390
#> [2,] 61.27946 -224.2951 -582.6267 255.5369 615.7415 827.1254 378.974 511.0628
#>         [,38]     [,39]     [,40]     [,41]      [,42]    [,43]    [,44]
#> [1,]  87.5784  52.23936 -154.2132 -150.1885  -81.25957 111.6271 17.95534
#> [2,] 164.7718 101.55988 -278.1722 -251.1689 -125.78373 227.6789 32.90389
#>          [,45]     [,46]    [,47]     [,48]    [,49]     [,50]     [,51]
#> [1,] -106.2325 -122.4789 132.1349  72.97289 208.7308 -63.25214 -25.92606
#> [2,] -196.1491 -232.3961 236.1802 128.84330 395.8592 -99.38362 -35.11113
#>          [,52]     [,53]     [,54]     [,55]     [,56]     [,57]     [,58]
#> [1,] -126.8758 -102.7317 -197.3345 -332.7755 -359.4376 -329.9048 -250.5564
#> [2,] -226.9220 -191.2606 -359.1300 -604.3822 -650.5624 -602.3997 -454.3023
#>          [,59]    [,60]     [,61]     [,62]     [,63]     [,64]     [,65]
#> [1,] -36.55789 23.01162 -268.9189 -393.2662 -424.8196 -251.2064 -22.37192
#> [2,] -58.40703 46.98156 -507.6626 -739.4053 -792.1252 -452.2700 -20.31455
#>         [,66]     [,67]      [,68]      [,69]      [,70]     [,71]    [,72]
#> [1,] 40.94632 -46.25241  -653.6914  -813.3189  -566.8655 -233.5401 10.96136
#> [2,] 87.04101 -87.76404 -1219.4810 -1503.8598 -1049.9575 -437.1713 23.51702
#>          [,73]     [,74]    [,75]    [,76]    [,77]    [,78]    [,79]    [,80]
#> [1,] -221.4051  68.95254 189.7447 415.4517 539.8777 466.3147 293.5784 153.1721
#> [2,] -414.7632 132.98857 349.9108 772.4732 998.9624 872.7065 536.2031 279.9630
#>          [,81]     [,82]     [,83]     [,84]    [,85]     [,86]     [,87]
#> [1,] -19.02569  76.30937 -142.3323 -250.6831 22.57060 -33.64186 -169.7312
#> [2,] -35.49724 163.50712 -250.5480 -454.4357 49.86685 -48.65536 -317.3872
#>          [,88]     [,89]    [,90]     [,91]    [,92]    [,93]    [,94]
#> [1,] -381.6736 -216.1095 225.8681  578.5261 457.7733 149.2432 232.1446
#> [2,] -712.0338 -420.2305 416.7541 1079.3880 855.8649 261.0236 428.2309
#>          [,95]     [,96]    [,97]     [,98]    [,99]   [,100]
#> [1,]  570.4336  601.9649 27.10469 -51.30105 28.64949 209.7399
#> [2,] 1062.6392 1121.9899 45.72570 -93.66679 45.42651 373.5156
p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_inv(p, data)
#>            [,1]      [,2]     [,3]     [,4]       [,5]          [,6]
#> [1,] -0.6276353 -41.48635 13286.46 -4465983 1500443601 -504109376285
#> [2,]  0.9230407 -71.14447 24516.98 -8235096 2766782460 -929565654859
#>              [,7]          [,8]         [,9]         [,10]        [,11]
#> [1,] 1.693674e+14 -5.690297e+16 1.911789e+19 -6.423106e+21 2.157993e+24
#> [2,] 3.123095e+14 -1.049277e+17 3.525294e+19 -1.184405e+22 3.979288e+24
#>              [,12]        [,13]         [,14]        [,15]         [,16]
#> [1,] -7.250287e+26 2.435905e+29 -8.183995e+31 2.749606e+34 -9.237947e+36
#> [2,] -1.336936e+27 4.491750e+29 -1.509109e+32 5.070207e+34 -1.703455e+37
#>             [,17]         [,18]        [,19]         [,20]        [,21]
#> [1,] 3.103706e+39 -1.042763e+42 3.503408e+44 -1.177052e+47 3.954584e+49
#> [2,] 5.723159e+39 -1.922830e+42 6.460200e+44 -2.170456e+47 7.292159e+49
#>              [,22]        [,23]         [,24]        [,25]        [,26]
#> [1,] -1.328636e+52 4.463864e+54 -1.499740e+57 5.038730e+59 -1.69288e+62
#> [2,] -2.449972e+52 8.231259e+54 -2.765485e+57 9.291298e+59 -3.12163e+62
#>             [,27]         [,28]        [,29]         [,30]        [,31]
#> [1,] 5.687629e+64 -1.910893e+67 6.420094e+69 -2.156982e+72 7.246888e+74
#> [2,] 1.048785e+65 -3.523641e+67 1.183850e+70 -3.977423e+72 1.336309e+75
#>              [,32]        [,33]         [,34]        [,35]         [,36]
#> [1,] -2.434762e+77 8.180158e+79 -2.748316e+82 9.233616e+84 -3.102251e+87
#> [2,] -4.489644e+77 1.508402e+80 -5.067830e+82 1.702657e+85 -5.720476e+87
#>             [,37]         [,38]        [,39]        [,40]         [,41]
#> [1,] 1.042274e+90 -3.501765e+92 1.176500e+95 -3.95273e+97 1.328013e+100
#> [2,] 1.921929e+90 -6.457171e+92 2.169439e+95 -7.28874e+97 2.448824e+100
#>               [,42]         [,43]          [,44]         [,45]          [,46]
#> [1,] -4.461771e+102 1.499037e+105 -5.036368e+107 1.692086e+110 -5.684962e+112
#> [2,] -8.227400e+102 2.764189e+105 -9.286942e+107 3.120167e+110 -1.048293e+113
#>              [,47]          [,48]         [,49]          [,50]         [,51]
#> [1,] 1.909997e+115 -6.417084e+117 2.155970e+120 -7.243490e+122 2.433621e+125
#> [2,] 3.521989e+115 -1.183295e+118 3.975558e+120 -1.335682e+123 4.487539e+125
#>               [,52]         [,53]          [,54]         [,55]          [,56]
#> [1,] -8.176322e+127 2.747028e+130 -9.229287e+132 3.100796e+135 -1.041785e+138
#> [2,] -1.507694e+128 5.065454e+130 -1.701858e+133 5.717794e+135 -1.921028e+138
#>              [,57]          [,58]         [,59]          [,60]         [,61]
#> [1,] 3.500123e+140 -1.175949e+143 3.950877e+145 -1.327390e+148 4.459679e+150
#> [2,] 6.454144e+140 -2.168422e+143 7.285323e+145 -2.447676e+148 8.223542e+150
#>               [,62]         [,63]          [,64]         [,65]          [,66]
#> [1,] -1.498334e+153 5.034007e+155 -1.691293e+158 5.682297e+160 -1.909101e+163
#> [2,] -2.762893e+153 9.282588e+155 -3.118704e+158 1.047802e+161 -3.520337e+163
#>              [,67]          [,68]         [,69]          [,70]         [,71]
#> [1,] 6.414075e+165 -2.154960e+168 7.240094e+170 -2.432480e+173 8.172489e+175
#> [2,] 1.182740e+166 -3.973694e+168 1.335056e+171 -4.485435e+173 1.506987e+176
#>               [,72]         [,73]          [,74]         [,75]          [,76]
#> [1,] -2.745740e+178 9.224960e+180 -3.099342e+183 1.041297e+186 -3.498482e+188
#> [2,] -5.063079e+178 1.701061e+181 -5.715113e+183 1.920127e+186 -6.451118e+188
#>              [,77]          [,78]         [,79]          [,80]         [,81]
#> [1,] 1.175398e+191 -3.949025e+193 1.326768e+196 -4.457589e+198 1.497632e+201
#> [2,] 2.167405e+191 -7.281907e+193 2.446528e+196 -8.219687e+198 2.761597e+201
#>               [,82]         [,83]          [,84]         [,85]          [,86]
#> [1,] -5.031646e+203 1.690500e+206 -5.679633e+208 1.908206e+211 -6.411068e+213
#> [2,] -9.278236e+203 3.117242e+206 -1.047311e+209 3.518687e+211 -1.182186e+214
#>              [,87]          [,88]         [,89]          [,90]         [,91]
#> [1,] 2.153949e+216 -7.236699e+218 2.431340e+221 -8.168657e+223 2.744453e+226
#> [2,] 3.971831e+216 -1.334430e+219 4.483332e+221 -1.506281e+224 5.060705e+226
#>               [,92]         [,93]          [,94]         [,95]          [,96]
#> [1,] -9.220635e+228 3.097889e+231 -1.040809e+234 3.496842e+236 -1.174846e+239
#> [2,] -1.700263e+229 5.712434e+231 -1.919227e+234 6.448093e+236 -2.166389e+239
#>              [,97]          [,98]         [,99]         [,100]
#> [1,] 3.947173e+241 -1.326146e+244 4.455499e+246 -1.496930e+249
#> [2,] 7.278493e+241 -2.445381e+244 8.215833e+246 -2.760302e+249