Outputs of an ARMA systems

This internal helper function computes the outputs of an ARMA system $$a_0 y_t + a_1 y_{t-1} + \cdots + a_p y_{t-p} = b_0 u_t + \cdots + b_q u_{t-q}$$

Usage

outputs_ARMA_cpp(A1, B, t0, u, y)

Arguments

A1

\((m, mp)\) matrix \(-a_0^{-1}(a_p,...,a_1)\).

B

\((m, n(q+1))\) matrix \(a_0^{-1}(b_0,...,b_q\).

t0

integer, start iteration at t = t0.

u

\((n, N)\) matrix with the inputs \((u_1,...,u_N\).

y

\((m, N)\) matrix with the outputs \((y_1,...,y_N\).

Value

This RcppArmadillo routine returns NULL but overwrites

the input argument y with the computed outputs!

Details

Values \(y_t\), \(u_t\) for \(t\leq 0\) are implicitly set to be zero. However, by starting the iteration with some \(t_0>1\) we can enforce non-zero initial values.

Note

Use this procedure with care!

The procedure does not check the input arguments. We require \(m > 0\), \(p \geq 0\), \(n(q+1) \geq 0\) and \(1 \leq t_0 \leq N\).

The procedure overwrites the input argument y.

The data matrices are organized columnwise (to avoid memory shuffling)!

Note also the non standard representation of the coefficient matrices.

See also

Examples

# generate a random ARMA(2,1) model (3 outputs, 2 inputs)
p = 2
q = 1
m = 3
n = 2
model = test_armamod(dim = c(m, n), degrees = c(p,q), digits = 2)
A = unclass(model$sys$a)
a0 = A[,,1]
A1 = -A[,,(p+1):2]
dim(A1) = c(m, m*p)
A1 = solve(a0, A1)
B = unclass(model$sys$b)
dim(B) = c(m, n*(q+1))
B = solve(a0, B)

# generate random noise sequence (sample size N = 10)
n.obs = 10
u = matrix(rnorm(n.obs*n), nrow = n, ncol = n.obs)
print(u)
#>            [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#> [1,] -0.6727197 -3.015827 0.3596626 -2.088264  0.2455572 0.9480790 2.1220845
#> [2,]  1.2258696  1.271059 0.6153130  1.785277 -1.3224404 0.5920432 0.6148116
#>            [,8]       [,9]      [,10]
#> [1,] 0.63043010 -1.2634735  1.5776279
#> [2,] 0.03644486 -0.4022728 -0.7037822

# generate matrix for the outputs
y = matrix(0, nrow = m, ncol = n.obs)

# call outputs_ARMA_cpp()
outputs_ARMA_cpp(A1, B, t0 = 2, u, y) # start with t>=2
print(u)
#>            [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#> [1,] -0.6727197 -3.015827 0.3596626 -2.088264  0.2455572 0.9480790 2.1220845
#> [2,]  1.2258696  1.271059 0.6153130  1.785277 -1.3224404 0.5920432 0.6148116
#>            [,8]       [,9]      [,10]
#> [1,] 0.63043010 -1.2634735  1.5776279
#> [2,] 0.03644486 -0.4022728 -0.7037822
print(y)  # y is overwritten with the computed outputs
#>      [,1]       [,2]      [,3]      [,4]      [,5]       [,6]       [,7]
#> [1,]    0 -2.0820766  3.020609 -8.863051 10.482240 -16.297199  27.933818
#> [2,]    0  1.7555846  0.750899  2.920361 -5.302189   6.870707 -11.521045
#> [3,]    0 -0.8398604 -3.843394 -2.204348 -2.866299  13.129438   2.798567
#>           [,8]      [,9]      [,10]
#> [1,] -39.19313  57.52580 -91.812319
#> [2,]  20.07121 -22.93972  42.908171
#> [3,]  13.97290 -29.83315   8.190574