Residuals of a statespace system

This internal helper function computes the residuals (and the directional derivatives of the residuals) for a statespace system of the form $$a_{t+1} = A a_t + B u_t, \; y_t = C a_t + D u_t$$ The system must be square and non-empty, i.e. \(m=n>0\).

Usage

residuals_STSP_cpp(A, B, C, D, y, a, u, dPI, dU)

Arguments

A

\((s,s)\) matrix.

B

\((s,m)\) matrix.

C

\((m,s)\) matrix.

D

\((m,m)\) matrix, must be regular.

y

\((m,N)\) matrix with the outputs: \((y_1,y_2,\ldots,y_N)\).

a

\((s,N+1)\) matrix. This matrix is overwritten with the (computed) states: \((a_1,a_2,\ldots,a_N,a_{N+1})\). On input a[,1] must hold the initial state \(a_1\).

u

\((m,N)\) matrix. This matrix is overwritten with (computed) residuals: \((u_1,u_2,\ldots,u_N)\).

dPI

\(((m+s)^2,K)\) matrix.

dU

\((mN,K)\) matrix or \((0,0)\) matrix. This matrix is overwritten with the directional derivatives of the residuals. However, if the matrix is empty then no derivatives are computed.

Value

This RcppArmadillo implementation returns NULL but overwrites the input arguments a, u and dU!

Note

Use this procedure with care!

The procedure does not check the input arguments.

The procedure overwrites the input arguments a, u and dU.

The data matrices are organized columnwise (to avoid memory shuffling)!

See also

Examples

# generate a random statespace model (3 outputs, 3 inputs and 4 states)
m = 2
s = 3
model = test_stspmod(dim = c(m, m), s = s, digits = 2)

# generate random noise sequence (sample size N = 10)
n.obs = 10
u = matrix(rnorm(n.obs*m), nrow = m, ncol = n.obs)

# generate matrix for the state sequence
a = matrix(0, nrow = s, ncol = n.obs+1)
a[,1] = rnorm(s) # random initial state a[1]

# generate matrix for the outputs
y = matrix(0, nrow = m, ncol = n.obs)

# compute outputs and states
outputs_STSP_cpp(model$sys$A, model$sys$B, model$sys$C, model$sys$D, u, a, y)

# recompute the states and disturbances/residuals from the given outputs:
uu = u + 0 # "deep copy" of the disturbances
aa = a + 0 # and the states
aa[, 2:(n.obs+1)] = 0 # clear all states a[t], t > 1
residuals_STSP_cpp(model$sys$A, model$sys$B, model$sys$C, model$sys$D,
                   y, aa, uu, diag(0), diag(0))
all.equal(u, uu) # check
#> [1] TRUE
all.equal(a, aa) # check
#> [1] TRUE

# compute directional derivatives of residuals
dPI = diag((m+s)^2)
dU = matrix(0, nrow = n.obs*m, ncol = ncol(dPI))
residuals_STSP_cpp(model$sys$A, model$sys$B, model$sys$C, model$sys$D,
                   y, aa, uu, dPI, dU)

# check the directional derivatives
eps = 1e-8
dU_num = matrix(0, nrow = m*n.obs, ncol = (m+s)^2)
dPI = matrix(0, nrow = (m+s), ncol = (m+s))
for (k in (1:((m+s)^2))) {
  dPI[] = 0
  dPI[k] = eps
  uu = u + 0 # "deep copy" of the disturbances
  aa = a + 0 # and the states
  residuals_STSP_cpp(model$sys$A + dPI[1:s,1:s],
                     model$sys$B + dPI[1:s,(s+1):(s+m)],
                     model$sys$C + dPI[(s+1):(s+m),1:s],
                     model$sys$D + dPI[(s+1):(s+m),(s+1):(s+m)],
                     y, aa, uu, diag(0), diag(0))
  dU_num[, k] = c(uu - u )/eps  # num. approx. of the derivative in direction "dPI"
}
# relative error of the numerical approximation
junk = (abs(dU)+abs(dU_num))
junk[junk == 0] = 1
2*abs(dU_num - dU)/junk
#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,] 2.000000e+00 2.000000e+00 2.000000e+00 8.200662e-08 2.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 5.412260e-10
#>  [3,] 1.974657e-07 2.353158e-06 2.737584e-07 3.519607e-09 1.286457e-08
#>  [4,] 1.694827e-07 1.148383e-07 2.546066e-07 1.206523e-07 3.918487e-09
#>  [5,] 3.155593e-08 3.124652e-07 8.299443e-08 9.116358e-09 5.370134e-08
#>  [6,] 4.259518e-08 6.798583e-09 1.195305e-09 5.513427e-08 1.674378e-08
#>  [7,] 2.909402e-08 1.152367e-08 6.479806e-08 2.537266e-08 6.273976e-08
#>  [8,] 4.470863e-09 3.588047e-08 6.560435e-07 5.339637e-09 1.953551e-08
#>  [9,] 3.410503e-08 3.309806e-06 6.579306e-08 2.405809e-08 7.874776e-08
#> [10,] 2.322148e-09 2.752168e-08 8.758034e-08 1.756672e-08 1.673058e-08
#> [11,] 3.708195e-08 2.141686e-07 6.479822e-08 2.252483e-08 1.172910e-07
#> [12,] 4.738577e-08 2.915763e-08 7.984212e-08 1.586337e-08 1.185376e-08
#> [13,] 3.721026e-08 1.676960e-07 6.424450e-08 2.096965e-08 1.407826e-07
#> [14,] 4.387813e-08 6.602925e-08 7.006413e-08 1.838257e-08 1.361328e-07
#> [15,] 3.685599e-08 1.403057e-07 6.731945e-08 2.283383e-08 1.462712e-07
#> [16,] 3.951300e-08 1.048041e-07 6.436786e-08 2.117752e-08 1.152099e-07
#> [17,] 3.916596e-08 1.344864e-07 6.882360e-08 2.253886e-08 1.590932e-07
#> [18,] 3.816934e-08 1.445894e-07 6.243880e-08 2.180462e-08 1.168610e-07
#> [19,] 3.825160e-08 1.316313e-07 6.988634e-08 2.342712e-08 1.553411e-07
#> [20,] 3.853148e-08 1.558423e-07 6.455501e-08 2.246559e-08 1.282321e-07
#>               [,6]         [,7]         [,8]         [,9]        [,10]
#>  [1,] 2.000000e+00 2.000000e+00 2.000000e+00 1.364860e-08 2.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 2.087975e-09
#>  [3,] 2.135569e-08 1.424199e-07 5.739534e-09 7.772962e-08 4.068000e-08
#>  [4,] 3.142091e-08 3.571074e-09 2.944101e-08 4.996770e-08 1.261012e-07
#>  [5,] 4.535335e-08 3.797182e-08 2.933679e-09 1.126135e-08 4.498874e-08
#>  [6,] 1.502630e-07 4.762111e-07 5.503734e-08 5.543286e-08 1.376049e-08
#>  [7,] 3.489132e-08 1.162124e-07 1.110105e-08 1.119964e-08 8.105491e-09
#>  [8,] 2.301001e-08 1.384923e-08 1.767969e-07 3.162220e-08 2.483838e-08
#>  [9,] 2.157367e-08 8.294491e-07 1.215169e-08 3.820112e-09 2.314601e-08
#> [10,] 9.403635e-09 1.314136e-08 5.045842e-09 1.490924e-08 9.598501e-09
#> [11,] 2.417651e-08 1.144960e-07 8.716369e-09 1.758141e-09 3.689531e-09
#> [12,] 4.779458e-08 1.436955e-08 3.648049e-09 5.559612e-09 1.742933e-09
#> [13,] 2.250698e-08 1.012662e-07 8.340884e-09 9.062887e-10 9.463784e-09
#> [14,] 3.920295e-08 7.482098e-09 4.202583e-09 2.827036e-09 1.257776e-08
#> [15,] 2.083945e-08 7.328468e-08 1.164660e-08 3.021842e-09 1.392137e-08
#> [16,] 3.458717e-08 2.505301e-08 4.987935e-09 4.071185e-09 8.004196e-08
#> [17,] 2.023426e-08 7.787487e-08 1.131059e-08 4.230896e-09 1.452783e-08
#> [18,] 3.286156e-08 3.174262e-08 6.012030e-09 4.130969e-09 1.364241e-08
#> [19,] 1.831711e-08 7.158990e-08 1.300912e-08 6.395691e-09 1.279239e-08
#> [20,] 3.179535e-08 4.470123e-08 7.449292e-09 5.320960e-09 9.498512e-09
#>              [,11]        [,12]        [,13]        [,14]        [,15]
#>  [1,] 2.000000e+00 2.000000e+00 2.000000e+00 2.113088e-08 2.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 2.113088e-08
#>  [3,] 3.384346e-08 3.498274e-07 3.763184e-08 1.822381e-08 9.868752e-08
#>  [4,] 2.352512e-08 2.426488e-09 8.095103e-08 4.283738e-09 3.610637e-08
#>  [5,] 1.424748e-06 3.838558e-08 4.381172e-08 1.026806e-08 1.197517e-08
#>  [6,] 3.279284e-07 4.424248e-08 7.240654e-08 1.901343e-08 3.439835e-08
#>  [7,] 1.847044e-07 6.357701e-08 7.143331e-08 1.076385e-08 4.511588e-08
#>  [8,] 1.241128e-08 1.916685e-09 7.507744e-09 1.142498e-08 1.003565e-08
#>  [9,] 1.446865e-07 3.973067e-08 7.701780e-08 1.636129e-08 4.626464e-08
#> [10,] 9.258457e-09 1.151009e-08 5.556689e-08 8.108560e-09 1.195220e-08
#> [11,] 1.520931e-07 7.217293e-08 4.437883e-08 7.505861e-09 6.997348e-08
#> [12,] 3.078456e-08 6.733252e-09 1.399157e-07 1.882991e-09 9.329321e-09
#> [13,] 1.550506e-07 1.878371e-07 4.018052e-08 6.194468e-09 1.031662e-07
#> [14,] 1.066989e-07 8.246083e-09 9.865285e-07 2.567540e-09 1.092047e-08
#> [15,] 1.567400e-07 6.391772e-07 4.376510e-08 4.453984e-09 1.198709e-07
#> [16,] 4.279680e-07 8.580580e-09 1.047771e-07 1.483980e-10 1.554924e-08
#> [17,] 1.608495e-07 1.968497e-07 3.983307e-08 4.071785e-09 1.652722e-07
#> [18,] 1.065341e-06 6.640607e-09 5.532773e-08 1.494455e-09 5.722388e-08
#> [19,] 1.610996e-07 1.704597e-07 4.731666e-08 2.437553e-09 1.725873e-07
#> [20,] 3.765076e-07 1.232736e-08 4.737720e-08 1.043289e-09 1.590234e-07
#>              [,16]        [,17]        [,18]        [,19]        [,20]
#>  [1,] 2.000000e+00 2.000000e+00 2.000000e+00 2.517430e-08 2.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 7.931119e-09
#>  [3,] 2.701349e-08 2.810508e-07 4.569277e-08 1.807897e-07 5.287432e-09
#>  [4,] 4.691518e-08 3.532497e-09 5.909404e-08 6.423297e-08 2.904479e-08
#>  [5,] 2.958046e-08 1.566616e-07 2.647735e-08 8.510074e-09 1.971775e-07
#>  [6,] 1.216080e-07 1.032995e-07 2.291298e-07 8.194030e-08 1.070052e-08
#>  [7,] 1.509118e-08 3.995125e-08 4.802400e-08 1.844399e-09 8.319462e-08
#>  [8,] 2.988354e-07 7.030607e-08 2.044729e-07 1.542820e-08 2.459689e-07
#>  [9,] 2.486663e-09 1.327880e-07 4.603175e-08 2.479323e-08 1.267753e-07
#> [10,] 7.081700e-08 1.408504e-07 7.807674e-08 4.436369e-08 4.945181e-08
#> [11,] 3.975434e-09 1.975663e-07 5.021211e-08 1.819272e-08 9.375013e-08
#> [12,] 3.242022e-08 5.863709e-08 7.532465e-08 2.688956e-08 1.643018e-07
#> [13,] 4.034405e-09 1.904318e-07 5.192739e-08 2.709207e-08 8.867315e-08
#> [14,] 2.807302e-08 1.466911e-07 5.599882e-08 2.856266e-08 9.988785e-08
#> [15,] 1.627931e-09 1.933216e-07 5.933045e-08 2.879202e-08 1.006784e-07
#> [16,] 2.116531e-08 1.266590e-07 4.795825e-08 2.663466e-08 6.171568e-08
#> [17,] 1.551984e-09 2.095134e-07 6.095434e-08 3.145126e-08 9.967690e-08
#> [18,] 2.079670e-08 1.594370e-07 4.990563e-08 2.836943e-08 5.947579e-08
#> [19,] 1.162058e-09 1.891414e-07 6.195757e-08 3.421496e-08 9.922121e-08
#> [20,] 1.900021e-08 1.745029e-07 5.366505e-08 3.078429e-08 6.270546e-08
#>              [,21]        [,22]        [,23]        [,24]        [,25]
#>  [1,] 2.000000e+00 2.000000e+00 2.000000e+00 3.391148e-08 2.000000e+00
#>  [2,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 3.391148e-08
#>  [3,] 4.757876e-08 1.230667e-07 5.130350e-08 3.269700e-09 4.050297e-08
#>  [4,] 4.746016e-08 8.710723e-09 7.219059e-08 1.147057e-07 2.478643e-08
#>  [5,] 2.097556e-08 1.014689e-07 1.852530e-08 1.501485e-08 1.677958e-08
#>  [6,] 1.070107e-09 2.523822e-08 6.990330e-08 4.335604e-08 9.108243e-08
#>  [7,] 1.788375e-08 7.737607e-08 2.875532e-08 2.091239e-08 6.348618e-10
#>  [8,] 6.245609e-09 2.490752e-08 9.245391e-08 7.622976e-09 3.368694e-08
#>  [9,] 1.968014e-08 1.321935e-07 2.694107e-08 2.112952e-08 3.362508e-08
#> [10,] 3.717839e-08 4.477251e-08 3.600429e-08 2.430098e-08 1.177394e-07
#> [11,] 2.067851e-08 5.187453e-08 2.821735e-08 2.379298e-08 3.351400e-08
#> [12,] 1.922321e-08 6.188499e-08 3.752173e-08 2.087111e-08 6.223610e-07
#> [13,] 1.995956e-08 6.129895e-08 2.997162e-08 2.409558e-08 3.915204e-08
#> [14,] 2.600901e-08 6.788867e-08 2.839683e-08 2.188247e-08 1.086934e-07
#> [15,] 1.759721e-08 6.384287e-08 3.012928e-08 2.689016e-08 4.697815e-08
#> [16,] 2.529412e-08 7.209754e-08 2.974592e-08 2.401967e-08 6.387716e-08
#> [17,] 1.851984e-08 5.138726e-08 3.003835e-08 2.611788e-08 4.188009e-08
#> [18,] 2.618624e-08 5.844055e-08 2.935291e-08 2.644662e-08 3.543420e-08
#> [19,] 1.626090e-08 6.073852e-08 3.187251e-08 2.805640e-08 4.594804e-08
#> [20,] 2.553554e-08 5.957578e-08 2.853028e-08 2.728535e-08 3.750185e-08