Constructor for RMFD Models

Usage

rmfdmod(sys, sigma_L = NULL, names = NULL, label = NULL)

Arguments

sys

rationalmatrices::rmfd() object

sigma_L

Left-factor of noise covariance, i.e. the covariance \(\sigma\) is obtained as sigma_L * t(sigma_L). If sigma_L is a vector of dimension \(n\), where \(n\) is the input dimension, only the diagonal elements are parametrized. If it is a vector of dimension \(n^2\), then the elements of sigma_L are filled column by column.

names

optional vector of character strings

label

optional character string

Value

Object of class rmfdmod.

Details

A right-matrix fraction description (RMFD) plus parameterisation of noise covariance. In (Hannan and Deistler 2012) , RMFDs are also called dynamic adjustment forms. Internally, MFDs are lists with slots sys, sigma_L, names, label. Many of the generic functions which construct derived objects like the autocovariance autocov() are not yet implemented for rmfdmod objects.

References

Hannan EJ, Deistler M (2012). The Statistical Theory of Linear Systems, Classics in Applied Mathematics. SIAM, Philadelphia. Originally published: John Wiley & Sons, New York, 1988.

Examples

y = rmfdmod(sys = test_rmfd(dim = c(3,2), degrees = c(2,2)))
y
#> RMFD model [3,2] with orders p = 2 and q = 2
#> right factor polynomial c(z):
#>      z^0 [,1]  [,2]   z^1 [,1]        [,2]  z^2 [,1]     [,2]
#> [1,]        1     0 -0.4649891  1.80524427 0.2090350 1.071341
#> [2,]        0     1  0.2814555 -0.03393583 0.2405301 1.227095
#> left factor polynomial d(z):
#>        z^0 [,1]       [,2]   z^1 [,1]       [,2]    z^2 [,1]       [,2]
#> [1,]  0.4044407  0.6731594  0.8905784 -1.3908101 -0.04089047 -1.3665964
#> [2,] -1.1248874 -0.2061704 -2.1587362  0.3613722 -0.45142477  1.3584698
#> [3,]  0.5185099  0.6854236 -2.2851640 -0.6967486  0.60229165  0.9553957
#> Left square root of noise covariance Sigma:
#>      u[1] u[2]
#> u[1]    1    0
#> u[2]    0    1