Toeplitz Calculations

Multiplication of stacked data vector with a block Toeplitz matrix (toepl_fwd() for MA calculations) or "inversion" of a block Toeplitz matrix in order to perform calculations equivalent to multiplying a given stacked data vector with the inverse of a lower-triangular banded block Toeplitz matrix (toepl_inv() for AR calculations). Note that matrix polynomials can be mapped one-to-one to banded lower-triangular block Toeplitz matrices.

Usage

toepl_fwd(polm_wide, data_in, t0 = 1)

toepl_inv(polm_rev, data_in, t0 = 1)

Arguments

polm_wide

Wide matrix \(( d_0, d_1, \ldots, d_q )\) of dimension \((m \times (q+1) n)\) which represents a matrix polynomial \(d(z)\) of degree \(q\).

data_in

Data matrix of dimension \((dim(inputs) x nobs)\). Every column corresponds to one observation.

t0

Integer. Time index where calculations should start. Default set to 1. For AR calculations, \(degree + 1\) would be another smart option.

polm_rev

Wide matrix \((c_p, ... , c_1)\) of dimension \((n \times (q+1) n)\) with coefficients ordered in reverse direction, and zero-lag coefficient matrix not included. It represents a square polynomial matrix \(c(z)\) with \(c_0\) equal to the identity matrix and of degree \(p\).

Value

data_out Data matrix of dimension \((dim_out x n_obs)\)

MA-type Toeplitz calculations

Given a polynomial matrix of degree \(q\) and dimension \((m \times n)\), where \(m \geq n\), and given a "wide" input data matrix of dimension \((n \times nobs)\), where \(nobs\) is the number of observations such that each column corresponds to one observation and the number of columns is equal to the number of observations, we calculate a "wide" output data matrix of dimension \((m \times nobs)\).

The function name toepl_fwd stems from the multiplication of the "stacked" input data vector \((u_1', \ldots , u_{nobs}')'\) with a banded lower-triangular block Toeplitz matrix \(T\) of dimension \((nobs m \times nobs n)\) whose block elements depend only on the difference between the row- and column-index such that \(T_{i,j} = d_{i-j}\).

AR-type Toeplitz calcuations

Given a square polynomial matrix \(c(z)\) for which \(c_0\) is the identity matrix and given a wide data matrix y = data_in, obtain the solution u, a wide data matrix, of the Toeplitz equation $$T (y_1', \ldots , y_{nobs}')' = (u_1', ... , u_{nobs}')'$$

Note that the zero-lag coefficient is discarded and the coefficients are in reversed order since this simplifies computations and implementation.

Examples

p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_fwd(p, data)
#>          [,1]      [,2]      [,3]       [,4]       [,5]      [,6]      [,7]
#> [1,] 141.4748 -23.13566 -248.4166  -591.9084  -590.1148 -267.8618 -22.54160
#> [2,] 265.6674 -40.86986 -473.6662 -1108.6499 -1114.6648 -508.4464 -40.41374
#>          [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#> [1,] 337.9961  61.65166 -174.3230 -480.3138 -332.2220 -359.2901 -468.8967
#> [2,] 633.4052 133.31731 -327.9794 -883.4771 -618.9747 -667.0883 -881.7640
#>          [,15]     [,16]     [,17]     [,18]     [,19]      [,20]     [,21]
#> [1,] -457.8742 -437.2916 -257.6943 -361.0563 -102.0771  -80.64441 -109.7963
#> [2,] -852.6561 -818.1317 -478.0183 -664.4814 -179.1595 -131.14669 -197.6166
#>         [,22]    [,23]     [,24]     [,25]     [,26]    [,27]    [,28]    [,29]
#> [1,] 34.91040 48.87798  86.59909  -93.8866 -292.7939 24.56999 130.1283 429.0934
#> [2,] 67.41063 89.49313 160.32566 -172.1910 -548.8772 40.10576 252.6989 809.0855
#>         [,30]     [,31]     [,32]     [,33]    [,34]    [,35]    [,36]
#> [1,] 222.2223 -12.03776 -56.45539  86.37366 305.3626 348.6875 152.3885
#> [2,] 431.3414 -16.82197 -90.71190 179.31578 586.9399 650.0350 271.0928
#>          [,37]     [,38]     [,39]    [,40]    [,41]    [,42]      [,43]
#> [1,] -11.71296 -25.25584  52.99139 315.1927 273.1658 31.19768  -60.09739
#> [2,] -26.60370 -41.56590 107.53906 588.6766 508.6323 57.19161 -118.66863
#>          [,44]    [,45]    [,46]    [,47]     [,48]      [,49]     [,50]
#> [1,] -189.3048 389.9989 175.8936 174.9174 -382.0507  -534.7276 -339.3511
#> [2,] -362.3038 708.3049 319.4822 316.9774 -718.7050 -1004.2183 -645.5925
#>          [,51]      [,52]     [,53]     [,54]     [,55]     [,56]     [,57]
#> [1,] -330.0379  -88.61569 -383.5234 -130.9149 -53.97294 -272.2546 -427.3010
#> [2,] -640.1781 -194.77295 -725.6863 -235.3338 -96.01573 -508.5599 -787.6784
#>          [,58]     [,59]      [,60]     [,61]     [,62]     [,63]      [,64]
#> [1,] -317.3687 -49.22824  -7.675721 -384.1177 -273.5994 -153.9431  -94.22858
#> [2,] -566.9174 -79.96722 -37.299727 -741.6514 -543.2314 -303.9576 -186.24470
#>          [,65]      [,66]     [,67]     [,68]    [,69]     [,70]    [,71]
#> [1,]  64.79213  -64.57484  62.96636 -39.36107 147.6626  87.72005 374.1570
#> [2,] 118.60555 -113.03190 111.50775 -76.14069 274.2275 170.02801 717.7904
#>          [,72]     [,73]     [,74]     [,75]     [,76]     [,77]     [,78]
#> [1,] -21.38605 -144.4953 -386.7786 -276.5288 -49.80268 -144.5885 -45.70775
#> [2,] -26.07724 -259.6862 -723.4931 -522.8544 -91.03230 -269.5783 -83.85605
#>          [,79]     [,80]     [,81]    [,82]    [,83]     [,84]     [,85]
#> [1,] -279.9757 -378.9133 -291.1977 127.3729 252.6703  93.75194 -191.5580
#> [2,] -522.4010 -707.0198 -540.1409 235.0461 474.4832 167.06251 -354.9276
#>          [,86]    [,87]    [,88]    [,89]     [,90]     [,91]    [,92]    [,93]
#> [1,] -319.0661 107.4696 185.0086 105.3729 -345.8812 -199.8807 129.1090 293.9800
#> [2,] -593.5456 213.2051 349.0362 187.9175 -659.1456 -386.8416 233.6103 542.2512
#>          [,94]       [,95]     [,96]     [,97]     [,98]     [,99]   [,100]
#> [1,]  58.82142 -5.46274106 -57.15317 -163.5273 -131.5317 -119.6936 42.34206
#> [2,] 108.85731 -0.06150599 -94.69699 -292.7382 -242.4070 -225.4060 75.00084
p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_inv(p, data)
#>           [,1]     [,2]       [,3]     [,4]         [,5]         [,6]
#> [1,] -0.145243 164.0438  -54728.88 18389019  -6178217732 2.075718e+12
#> [2,] -1.210946 299.3382 -100931.57 33908862 -11392486213 3.827574e+12
#>               [,7]         [,8]          [,9]        [,10]         [,11]
#> [1,] -6.973862e+14 2.343033e+17 -7.871971e+19 2.644774e+22 -8.885741e+24
#> [2,] -1.285964e+15 4.320497e+17 -1.451573e+20 4.876900e+22 -1.638509e+25
#>             [,12]         [,13]        [,14]         [,15]        [,16]
#> [1,] 2.985374e+27 -1.003007e+30 3.369837e+32 -1.132176e+35 3.803812e+37
#> [2,] 5.504958e+27 -1.849520e+30 6.213898e+32 -2.087705e+35 7.014138e+37
#>              [,17]        [,18]         [,19]        [,20]         [,21]
#> [1,] -1.277980e+40 4.293675e+42 -1.442561e+45 4.846623e+47 -1.628337e+50
#> [2,] -2.356565e+40 7.917433e+42 -2.660048e+45 8.937058e+47 -3.002615e+50
#>             [,22]         [,23]        [,24]         [,25]        [,26]
#> [1,] 5.470782e+52 -1.838038e+55 6.175321e+57 -2.074744e+60 6.970592e+62
#> [2,] 1.008799e+53 -3.389298e+55 1.138714e+58 -3.825779e+60 1.285361e+63
#>              [,27]        [,28]         [,29]        [,30]         [,31]
#> [1,] -2.341935e+65 7.868280e+67 -2.643534e+70 8.881575e+72 -2.983974e+75
#> [2,] -4.318471e+65 1.450892e+68 -4.874614e+70 1.637741e+73 -5.502377e+75
#>             [,32]         [,33]        [,34]         [,35]        [,36]
#> [1,] 1.002536e+78 -3.368257e+80 1.131645e+83 -3.802028e+85 1.277381e+88
#> [2,] 1.848653e+78 -6.210985e+80 2.086727e+83 -7.010849e+85 2.355460e+88
#>              [,37]        [,38]         [,39]        [,40]          [,41]
#> [1,] -4.291662e+90 1.441885e+93 -4.844351e+95 1.627574e+98 -5.468217e+100
#> [2,] -7.913721e+90 2.658801e+93 -8.932868e+95 3.001207e+98 -1.008326e+101
#>              [,42]          [,43]         [,44]          [,45]         [,46]
#> [1,] 1.837176e+103 -6.172426e+105 2.073772e+108 -6.967324e+110 2.340837e+113
#> [2,] 3.387709e+103 -1.138181e+106 3.823986e+108 -1.284758e+111 4.316447e+113
#>               [,47]         [,48]          [,49]         [,50]          [,51]
#> [1,] -7.864591e+115 2.642295e+118 -8.877411e+120 2.982575e+123 -1.002066e+126
#> [2,] -1.450212e+116 4.872328e+118 -1.636973e+121 5.499797e+123 -1.847786e+126
#>              [,52]          [,53]         [,54]          [,55]         [,56]
#> [1,] 3.366678e+128 -1.131115e+131 3.800246e+133 -1.276782e+136 4.289649e+138
#> [2,] 6.208073e+128 -2.085748e+131 7.007562e+133 -2.354355e+136 7.910011e+138
#>               [,57]        [,58]          [,59]         [,60]          [,61]
#> [1,] -1.441209e+141 4.84208e+143 -1.626811e+146 5.465653e+148 -1.836315e+151
#> [2,] -2.657554e+141 8.92868e+143 -2.999800e+146 1.007853e+149 -3.386121e+151
#>              [,62]          [,63]         [,64]          [,65]         [,66]
#> [1,] 6.169532e+153 -2.072799e+156 6.964058e+158 -2.339739e+161 7.860904e+163
#> [2,] 1.137647e+154 -3.822193e+156 1.284156e+159 -4.314423e+161 1.449532e+164
#>               [,67]         [,68]          [,69]         [,70]          [,71]
#> [1,] -2.641056e+166 8.873248e+168 -2.981177e+171 1.001596e+174 -3.365099e+176
#> [2,] -4.870044e+166 1.636206e+169 -5.497219e+171 1.846920e+174 -6.205162e+176
#>              [,72]          [,73]         [,74]          [,75]         [,76]
#> [1,] 1.130584e+179 -3.798464e+181 1.276183e+184 -4.287638e+186 1.440533e+189
#> [2,] 2.084770e+179 -7.004277e+181 2.353252e+184 -7.906303e+186 2.656308e+189
#>               [,77]         [,78]          [,79]         [,80]          [,81]
#> [1,] -4.839810e+191 1.626048e+194 -5.463091e+196 1.835454e+199 -6.166639e+201
#> [2,] -8.924493e+191 2.998394e+194 -1.007381e+197 3.384533e+199 -1.137114e+202
#>              [,82]          [,83]         [,84]          [,85]         [,86]
#> [1,] 2.071828e+204 -6.960793e+206 2.338642e+209 -7.857219e+211 2.639818e+214
#> [2,] 3.820401e+204 -1.283554e+207 4.312400e+209 -1.448852e+212 4.867760e+214
#>               [,87]         [,88]          [,89]         [,90]          [,91]
#> [1,] -8.869088e+216 2.979779e+219 -1.001127e+222 3.363521e+224 -1.130054e+227
#> [2,] -1.635439e+217 5.494641e+219 -1.846054e+222 6.202253e+224 -2.083793e+227
#>              [,92]          [,93]         [,94]          [,95]         [,96]
#> [1,] 3.796683e+229 -1.275585e+232 4.285628e+234 -1.439858e+237 4.837540e+239
#> [2,] 7.000993e+229 -2.352148e+232 7.902596e+234 -2.655063e+237 8.920309e+239
#>               [,97]         [,98]          [,99]        [,100]
#> [1,] -1.625286e+242 5.460529e+244 -1.834593e+247 6.163748e+249
#> [2,] -2.996988e+242 1.006909e+245 -3.382946e+247 1.136580e+250