Toeplitz Calculations

Multiplication of stacked data vector with a block Toeplitz matrix (toepl_fwd() for MA calculations) or "inversion" of a block Toeplitz matrix in order to perform calculations equivalent to multiplying a given stacked data vector with the inverse of a lower-triangular banded block Toeplitz matrix (toepl_inv() for AR calculations). Note that matrix polynomials can be mapped one-to-one to banded lower-triangular block Toeplitz matrices.

Usage

toepl_fwd(polm_wide, data_in, t0 = 1)

toepl_inv(polm_rev, data_in, t0 = 1)

Arguments

polm_wide

Wide matrix \(( d_0, d_1, \ldots, d_q )\) of dimension \((m \times (q+1) n)\) which represents a matrix polynomial \(d(z)\) of degree \(q\).

data_in

Data matrix of dimension \((dim(inputs) x nobs)\). Every column corresponds to one observation.

t0

Integer. Time index where calculations should start. Default set to 1. For AR calculations, \(degree + 1\) would be another smart option.

polm_rev

Wide matrix \((c_p, ... , c_1)\) of dimension \((n \times (q+1) n)\) with coefficients ordered in reverse direction, and zero-lag coefficient matrix not included. It represents a square polynomial matrix \(c(z)\) with \(c_0\) equal to the identity matrix and of degree \(p\).

Value

data_out Data matrix of dimension \((dim_out x n_obs)\)

MA-type Toeplitz calculations

Given a polynomial matrix of degree \(q\) and dimension \((m \times n)\), where \(m \geq n\), and given a "wide" input data matrix of dimension \((n \times nobs)\), where \(nobs\) is the number of observations such that each column corresponds to one observation and the number of columns is equal to the number of observations, we calculate a "wide" output data matrix of dimension \((m \times nobs)\).

The function name toepl_fwd stems from the multiplication of the "stacked" input data vector \((u_1', \ldots , u_{nobs}')'\) with a banded lower-triangular block Toeplitz matrix \(T\) of dimension \((nobs m \times nobs n)\) whose block elements depend only on the difference between the row- and column-index such that \(T_{i,j} = d_{i-j}\).

AR-type Toeplitz calcuations

Given a square polynomial matrix \(c(z)\) for which \(c_0\) is the identity matrix and given a wide data matrix y = data_in, obtain the solution u, a wide data matrix, of the Toeplitz equation $$T (y_1', \ldots , y_{nobs}')' = (u_1', ... , u_{nobs}')'$$

Note that the zero-lag coefficient is discarded and the coefficients are in reversed order since this simplifies computations and implementation.

Examples

p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_fwd(p, data)
#>           [,1]       [,2]       [,3]     [,4]      [,5]     [,6]      [,7]
#> [1,] -230.4269  -65.24943  -96.21957 227.1020  88.04089 143.0543 -54.29231
#> [2,] -427.8741 -115.55100 -172.53560 417.8549 161.21812 263.2176 -87.74698
#>            [,8]      [,9]    [,10]    [,11]     [,12]    [,13]     [,14]
#> [1,]  -87.36762 -37.50433 135.9921 383.1742  610.9087  840.138  654.7003
#> [2,] -150.81496 -51.12160 249.1286 723.7437 1134.5599 1568.011 1200.5716
#>         [,15]    [,16]    [,17]    [,18]    [,19]     [,20]      [,21]
#> [1,] 399.1284 250.2904 463.7450 348.4572 251.2799 -228.1338  -53.22238
#> [2,] 738.9657 470.4546 876.0141 648.3634 463.0158 -424.1497 -100.22835
#>           [,22]    [,23]    [,24]    [,25]     [,26]     [,27]      [,28]
#> [1,]  -84.61805 402.7743 289.2047 368.6538 -111.5608 -306.8564  -732.8137
#> [2,] -165.83320 729.1268 509.8431 663.0719 -218.3311 -571.3351 -1364.4290
#>          [,29]     [,30]    [,31]    [,32]    [,33]    [,34]     [,35]
#> [1,] -427.3068 -252.7918 140.2372 179.6362 241.5349 179.3384 -25.20387
#> [2,] -795.7710 -461.3084 273.6168 345.0547 458.6718 343.7640 -31.08292
#>          [,36]     [,37]     [,38]    [,39]     [,40]     [,41]      [,42]
#> [1,] -168.6850 -181.7739 -132.1682 29.62087 -234.8291 -181.9991  -4.215969
#> [2,] -308.9189 -335.0708 -240.5506 54.78590 -437.8064 -354.4975 -17.068119
#>          [,43]      [,44]      [,45]     [,46]     [,47]     [,48]     [,49]
#> [1,] -279.3680  -592.6477  -818.5562 -342.5250 -176.0954 -341.5443 -486.2283
#> [2,] -553.9136 -1120.9842 -1537.3071 -642.8229 -331.7771 -635.1541 -903.3418
#>          [,50]      [,51]     [,52]      [,53]     [,54]     [,55]    [,56]
#> [1,] -441.9474  -540.9005 -483.5068  -554.8233 -139.4079  54.96059 163.0957
#> [2,] -818.3608 -1003.6984 -903.1601 -1038.3572 -257.2232 112.11148 311.2134
#>          [,57]     [,58]     [,59]    [,60]    [,61]    [,62]     [,63]
#> [1,] -111.9891 -266.9240 -30.41314 239.6962 227.1011 14.25958 -192.8079
#> [2,] -209.5836 -496.7485 -54.67421 462.3486 426.3327 24.22993 -368.5653
#>         [,64]    [,65]    [,66]    [,67]    [,68]     [,69]    [,70]    [,71]
#> [1,] 21.58261 241.1887 411.3266 481.5223 232.7041  98.63027 11.69385 232.8797
#> [2,] 26.19979 441.3663 762.2428 892.3764 431.7609 171.19246 11.59549 418.5329
#>         [,72]    [,73]     [,74]    [,75]    [,76]    [,77]    [,78]    [,79]
#> [1,] 239.7535 342.2693  91.34935 144.8746 223.3980 495.8465 482.5913 328.7273
#> [2,] 436.3905 637.8153 175.63951 259.7936 401.3858 905.4047 885.4123 597.0252
#>          [,80]     [,81]     [,82]     [,83]     [,84]     [,85]    [,86]
#> [1,]  -48.5708 -346.4565  -549.851 -520.8243 -140.3527 -10.72253 412.6746
#> [2,] -101.2307 -655.9602 -1031.470 -974.2244 -261.7713 -25.80970 755.5504
#>         [,87]    [,88]     [,89]     [,90]    [,91]    [,92]    [,93]    [,94]
#> [1,] 285.2892 282.9047  78.94515 -102.1215 25.61300 21.28791 179.1452 174.9079
#> [2,] 520.4904 525.1103 150.53439 -194.2331 38.60141 35.09982 325.6536 314.8574
#>          [,95]   [,96]     [,97]    [,98]     [,99]    [,100]
#> [1,] -215.8110 2.72159 -204.5713 120.2459 -129.2890 -131.3231
#> [2,] -413.1627 9.06002 -375.6820 221.0298 -251.4689 -248.5277
p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_inv(p, data)
#>             [,1]      [,2]      [,3]     [,4]       [,5]          [,6]
#> [1,]  0.22058305 -21.44662  7051.521 -2369461  796074739 -267460050623
#> [2,] -0.03755653 -38.27895 13005.590 -4369218 1467942860 -493189948461
#>              [,7]          [,8]         [,9]         [,10]        [,11]
#> [1,] 8.985950e+13 -3.019041e+16 1.014318e+19 -3.407840e+21 1.144944e+24
#> [2,] 1.656988e+14 -5.567040e+16 1.870378e+19 -6.283976e+21 2.111250e+24
#>              [,12]        [,13]         [,14]        [,15]         [,16]
#> [1,] -3.846709e+26 1.292392e+29 -4.342097e+31 1.458830e+34 -4.901281e+36
#> [2,] -7.093240e+26 2.383141e+29 -8.006723e+31 2.690047e+34 -9.037846e+36
#>             [,17]         [,18]        [,19]         [,20]        [,21]
#> [1,] 1.646701e+39 -5.532479e+41 1.858766e+44 -6.244964e+46 2.098142e+49
#> [2,] 3.036477e+39 -1.020176e+42 3.427521e+44 -1.151556e+47 3.868925e+49
#>              [,22]        [,23]         [,24]        [,25]         [,26]
#> [1,] -7.049203e+51 2.368346e+54 -7.957015e+56 2.673347e+59 -8.981737e+61
#> [2,] -1.299856e+52 4.367173e+54 -1.467255e+57 4.929587e+59 -1.656211e+62
#>             [,27]         [,28]        [,29]         [,30]        [,31]
#> [1,] 3.017626e+64 -1.013842e+67 3.406242e+69 -1.144407e+72 3.844906e+74
#> [2,] 5.564430e+64 -1.869501e+67 6.281030e+69 -2.110260e+72 7.089914e+74
#>              [,32]        [,33]         [,34]        [,35]         [,36]
#> [1,] -1.291787e+77 4.340061e+79 -1.458146e+82 4.898983e+84 -1.645929e+87
#> [2,] -2.382024e+77 8.002969e+79 -2.688786e+82 9.033608e+84 -3.035053e+87
#>             [,37]         [,38]        [,39]         [,40]         [,41]
#> [1,] 5.529885e+89 -1.857895e+92 6.242036e+94 -2.097159e+97  7.045898e+99
#> [2,] 1.019698e+90 -3.425914e+92 1.151016e+95 -3.867111e+97 1.299247e+100
#>               [,42]         [,43]          [,44]         [,45]          [,46]
#> [1,] -2.367235e+102 7.953285e+104 -2.672093e+107 8.977526e+109 -3.016211e+112
#> [2,] -4.365126e+102 1.466567e+105 -4.927276e+107 1.655434e+110 -5.561821e+112
#>              [,47]          [,48]         [,49]          [,50]         [,51]
#> [1,] 1.013367e+115 -3.404645e+117 1.143871e+120 -3.843103e+122 1.291181e+125
#> [2,] 1.868625e+115 -6.278085e+117 2.109270e+120 -7.086590e+122 2.380907e+125
#>               [,52]         [,53]          [,54]         [,55]          [,56]
#> [1,] -4.338026e+127 1.457462e+130 -4.896686e+132 1.645157e+135 -5.527292e+137
#> [2,] -7.999216e+127 2.687525e+130 -9.029373e+132 3.033630e+135 -1.019220e+138
#>              [,57]          [,58]         [,59]          [,60]         [,61]
#> [1,] 1.857024e+140 -6.239109e+142 2.096176e+145 -7.042595e+147 2.366126e+150
#> [2,] 3.424308e+140 -1.150477e+143 3.865298e+145 -1.298638e+148 4.363079e+150
#>               [,62]         [,63]          [,64]         [,65]          [,66]
#> [1,] -7.949556e+152 2.670840e+155 -8.973317e+157 3.014797e+160 -1.012892e+163
#> [2,] -1.465879e+153 4.924966e+155 -1.654658e+158 5.559214e+160 -1.867749e+163
#>              [,67]          [,68]         [,69]          [,70]         [,71]
#> [1,] 3.403049e+165 -1.143334e+168 3.841301e+170 -1.290576e+173 4.335992e+175
#> [2,] 6.275141e+165 -2.108281e+168 7.083268e+170 -2.379790e+173 7.995466e+175
#>               [,72]         [,73]          [,74]         [,75]          [,76]
#> [1,] -1.456779e+178 4.894391e+180 -1.644386e+183 5.524701e+185 -1.856153e+188
#> [2,] -2.686265e+178 9.025140e+180 -3.032208e+183 1.018742e+186 -3.422702e+188
#>              [,77]          [,78]         [,79]          [,80]         [,81]
#> [1,] 6.236184e+190 -2.095193e+193 7.039293e+195 -2.365016e+198 7.945829e+200
#> [2,] 1.149937e+191 -3.863485e+193 1.298029e+196 -4.361033e+198 1.465192e+201
#>               [,82]         [,83]          [,84]         [,85]          [,86]
#> [1,] -2.669588e+203 8.969110e+205 -3.013383e+208 1.012417e+211 -3.401454e+213
#> [2,] -4.922657e+203 1.653882e+206 -5.556607e+208 1.866873e+211 -6.272199e+213
#>              [,87]          [,88]         [,89]          [,90]         [,91]
#> [1,] 1.142798e+216 -3.839500e+218 1.289970e+221 -4.333959e+223 1.456096e+226
#> [2,] 2.107293e+216 -7.079947e+218 2.378675e+221 -7.991717e+223 2.685006e+226
#>               [,92]         [,93]          [,94]         [,95]          [,96]
#> [1,] -4.892096e+228 1.643615e+231 -5.522111e+233 1.855283e+236 -6.233260e+238
#> [2,] -9.020908e+228 3.030786e+231 -1.018264e+234 3.421098e+236 -1.149398e+239
#>              [,97]          [,98]         [,99]         [,100]
#> [1,] 2.094210e+241 -7.035993e+243 2.363907e+246 -7.942103e+248
#> [2,] 3.861674e+241 -1.297420e+244 4.358989e+246 -1.464505e+249