Toeplitz Calculations

Multiplication of stacked data vector with a block Toeplitz matrix (toepl_fwd() for MA calculations) or "inversion" of a block Toeplitz matrix in order to perform calculations equivalent to multiplying a given stacked data vector with the inverse of a lower-triangular banded block Toeplitz matrix (toepl_inv() for AR calculations). Note that matrix polynomials can be mapped one-to-one to banded lower-triangular block Toeplitz matrices.

Usage

toepl_fwd(polm_wide, data_in, t0 = 1)

toepl_inv(polm_rev, data_in, t0 = 1)

Arguments

polm_wide

Wide matrix \(( d_0, d_1, \ldots, d_q )\) of dimension \((m \times (q+1) n)\) which represents a matrix polynomial \(d(z)\) of degree \(q\).

data_in

Data matrix of dimension \((dim(inputs) x nobs)\). Every column corresponds to one observation.

t0

Integer. Time index where calculations should start. Default set to 1. For AR calculations, \(degree + 1\) would be another smart option.

polm_rev

Wide matrix \((c_p, ... , c_1)\) of dimension \((n \times (q+1) n)\) with coefficients ordered in reverse direction, and zero-lag coefficient matrix not included. It represents a square polynomial matrix \(c(z)\) with \(c_0\) equal to the identity matrix and of degree \(p\).

Value

data_out Data matrix of dimension \((dim_out x n_obs)\)

MA-type Toeplitz calculations

Given a polynomial matrix of degree \(q\) and dimension \((m \times n)\), where \(m \geq n\), and given a "wide" input data matrix of dimension \((n \times nobs)\), where \(nobs\) is the number of observations such that each column corresponds to one observation and the number of columns is equal to the number of observations, we calculate a "wide" output data matrix of dimension \((m \times nobs)\).

The function name toepl_fwd stems from the multiplication of the "stacked" input data vector \((u_1', \ldots , u_{nobs}')'\) with a banded lower-triangular block Toeplitz matrix \(T\) of dimension \((nobs m \times nobs n)\) whose block elements depend only on the difference between the row- and column-index such that \(T_{i,j} = d_{i-j}\).

AR-type Toeplitz calcuations

Given a square polynomial matrix \(c(z)\) for which \(c_0\) is the identity matrix and given a wide data matrix y = data_in, obtain the solution u, a wide data matrix, of the Toeplitz equation $$T (y_1', \ldots , y_{nobs}')' = (u_1', ... , u_{nobs}')'$$

Note that the zero-lag coefficient is discarded and the coefficients are in reversed order since this simplifies computations and implementation.

Examples

p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_fwd(p, data)
#>           [,1]      [,2]     [,3]      [,4]     [,5]      [,6]      [,7]
#> [1,]  77.88936 -141.7628 145.1943 -100.9090  86.6346 -327.4640 -161.8148
#> [2,] 151.12709 -246.9925 299.7344 -159.5419 166.9235 -611.6294 -312.3931
#>           [,8]      [,9]      [,10]     [,11]     [,12]     [,13]    [,14]
#> [1,] -122.2955 -117.6364  -82.15673 -199.9383 -192.9253 -109.3355 21.19757
#> [2,] -232.5138 -236.5170 -164.31963 -383.3160 -364.9630 -213.8007 14.69428
#>         [,15]     [,16]     [,17]      [,18]    [,19]     [,20]     [,21]
#> [1,] 14.74409 -159.3111 -126.7744  -84.50766 184.4034 -17.82726 -114.0374
#> [2,] 13.31315 -307.4891 -227.0016 -155.71573 347.8717 -35.51716 -224.6815
#>          [,22]     [,23]     [,24]     [,25]     [,26]    [,27]    [,28]
#> [1,] -503.8494 -324.5687 -160.9731  80.85545 -62.59665 34.80536 47.63224
#> [2,] -951.0216 -614.2522 -296.6608 161.43446 -90.88371 77.15831 84.14289
#>         [,29]    [,30]    [,31]    [,32]      [,33]     [,34]     [,35]
#> [1,] 239.3154 397.2031 513.1338 459.5870  -68.69651 -378.6913 -228.7658
#> [2,] 430.7602 726.5041 960.0394 851.9137 -128.20735 -700.4899 -421.1819
#>         [,36]    [,37]      [,38]     [,39]    [,40]     [,41]    [,42]
#> [1,] 28.19113 20.06021  -91.16675 -32.44854 101.4258  95.15421 114.0035
#> [2,] 64.74845 47.13291 -165.10154 -69.85698 171.4146 171.28522 211.6225
#>         [,43]    [,44]    [,45]    [,46]    [,47]    [,48]    [,49]    [,50]
#> [1,] 132.0305 190.7212 352.5188 541.2942 461.5141 320.0934 218.2515 298.9156
#> [2,] 241.6069 344.3757 645.2971 987.7103 825.7957 562.3157 389.7961 566.0852
#>         [,51]    [,52]     [,53]     [,54]    [,55]    [,56]    [,57]    [,58]
#> [1,] 390.2281 199.7450  74.57167  4.908045 230.9796 478.5157 413.8979 43.84532
#> [2,] 737.7278 391.4046 166.36878 37.112500 439.5847 884.1896 763.1475 78.26082
#>          [,59]     [,60]     [,61]    [,62]     [,63]    [,64]    [,65]
#> [1,] -246.3377 -301.4791 -154.6170 44.41441  89.52473 201.5594 130.0858
#> [2,] -459.7251 -551.6215 -279.5503 91.20776 168.38790 386.4406 253.9928
#>         [,66]     [,67]     [,68]    [,69]    [,70]    [,71]      [,72]
#> [1,] 201.4582 -38.20403  3.467639 23.75060 433.8254 149.8919  -75.02994
#> [2,] 371.2796 -88.80823 -1.339687 54.97751 822.2681 288.1637 -142.56871
#>           [,73]    [,74]    [,75]    [,76]    [,77]     [,78]     [,79]
#> [1,] -1.4489556 154.6870 462.4494 207.6014 207.0852 -190.6689 -276.8920
#> [2,]  0.6115891 274.5521 861.9501 390.5182 404.4998 -346.6210 -517.7392
#>          [,80]    [,81]     [,82]     [,83]      [,84]    [,85]    [,86]
#> [1,] -348.5044 11.44899 -33.07494 -38.59866  -80.15798 215.7784 226.3083
#> [2,] -653.1656 21.09577 -63.10823 -87.14130 -151.63902 401.5095 430.7965
#>          [,87]      [,88]     [,89]      [,90]    [,91]    [,92]     [,93]
#> [1,] -226.3582  -818.0346  -859.661  -543.4951 16.67412 103.7158  72.47077
#> [2,] -428.6238 -1515.1996 -1582.392 -1006.1654 26.66353 175.6057 133.84706
#>          [,94]      [,95]     [,96]    [,97]     [,98]     [,99]    [,100]
#> [1,] -197.7723  -87.75489  99.08135 148.6724  57.39691 -190.3617 -233.9043
#> [2,] -363.2818 -146.76630 201.76647 289.8853 105.65634 -370.1016 -452.2675
p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_inv(p, data)
#>            [,1]     [,2]       [,3]     [,4]         [,5]         [,6]
#> [1,]  0.2487596 200.9781  -66541.11 22359602  -7512220066 2.523907e+12
#> [2,] -1.8659559 362.5346 -122729.31 41230454 -13852354938 4.654025e+12
#>               [,7]         [,8]          [,9]        [,10]         [,11]
#> [1,] -8.479660e+14 2.848941e+17 -9.571689e+19 3.215834e+22 -1.080435e+25
#> [2,] -1.563629e+15 5.253380e+17 -1.764997e+20 5.929922e+22 -1.992297e+25
#>             [,12]         [,13]        [,14]         [,15]        [,16]
#> [1,] 3.629976e+27 -1.219576e+30 4.097453e+32 -1.376636e+35 4.625131e+37
#> [2,] 6.693590e+27 -2.248869e+30 7.555605e+32 -2.538483e+35 8.528632e+37
#>              [,17]        [,18]         [,19]        [,20]         [,21]
#> [1,] -1.553922e+40 5.220766e+42 -1.754039e+45 5.893108e+47 -1.979928e+50
#> [2,] -2.865395e+40 9.626968e+42 -3.234406e+45 1.086675e+48 -3.650940e+50
#>             [,22]         [,23]        [,24]         [,25]        [,26]
#> [1,] 6.652035e+52 -2.234908e+55 7.508698e+57 -2.522724e+60 8.475685e+62
#> [2,] 1.226619e+53 -4.121116e+55 1.384586e+58 -4.651843e+60 1.562896e+63
#>              [,27]        [,28]         [,29]        [,30]         [,31]
#> [1,] -2.847606e+65 9.567202e+67 -3.214327e+70 1.079929e+73 -3.628275e+75
#> [2,] -5.250917e+65 1.764169e+68 -5.927141e+70 1.991363e+73 -6.690452e+75
#>             [,32]         [,33]        [,34]         [,35]        [,36]
#> [1,] 1.219004e+78 -4.095532e+80 1.375990e+83 -4.622963e+85 1.553193e+88
#> [2,] 2.247815e+78 -7.552062e+80 2.537293e+83 -8.524633e+85 2.864051e+88
#>              [,37]        [,38]         [,39]        [,40]          [,41]
#> [1,] -5.218318e+90 1.753217e+93 -5.890345e+95 1.979000e+98 -6.648916e+100
#> [2,] -9.622454e+90 3.232890e+93 -1.086165e+96 3.649229e+98 -1.226044e+101
#>              [,42]          [,43]         [,44]          [,45]         [,46]
#> [1,] 2.233860e+103 -7.505178e+105 2.521541e+108 -8.471711e+110 2.846271e+113
#> [2,] 4.119184e+103 -1.383937e+106 4.649662e+108 -1.562163e+111 5.248455e+113
#>               [,47]         [,48]          [,49]         [,50]          [,51]
#> [1,] -9.562716e+115 3.212820e+118 -1.079422e+121 3.626573e+123 -1.218433e+126
#> [2,] -1.763342e+116 5.924363e+118 -1.990429e+121 6.687315e+123 -2.246761e+126
#>              [,52]          [,53]         [,54]          [,55]         [,56]
#> [1,] 4.093611e+128 -1.375345e+131 4.620796e+133 -1.552465e+136 5.215872e+138
#> [2,] 7.548522e+128 -2.536104e+131 8.520637e+133 -2.862708e+136 9.617943e+138
#>               [,57]         [,58]          [,59]         [,60]          [,61]
#> [1,] -1.752395e+141 5.887583e+143 -1.978072e+146 6.645799e+148 -2.232812e+151
#> [2,] -3.231374e+141 1.085656e+144 -3.647518e+146 1.225469e+149 -4.117253e+151
#>              [,62]          [,63]         [,64]          [,65]         [,66]
#> [1,] 7.501659e+153 -2.520359e+156 8.467739e+158 -2.844936e+161 9.558233e+163
#> [2,] 1.383288e+154 -4.647482e+156 1.561431e+159 -5.245994e+161 1.762515e+164
#>               [,67]         [,68]          [,69]         [,70]          [,71]
#> [1,] -3.211313e+166 1.078916e+169 -3.624873e+171 1.217861e+174 -4.091692e+176
#> [2,] -5.921585e+166 1.989496e+169 -6.684180e+171 2.245707e+174 -7.544983e+176
#>              [,72]          [,73]         [,74]          [,75]         [,76]
#> [1,] 1.374700e+179 -4.618629e+181 1.551737e+184 -5.213426e+186 1.751573e+189
#> [2,] 2.534914e+179 -8.516642e+181 2.861366e+184 -9.613433e+186 3.229859e+189
#>               [,77]         [,78]          [,79]         [,80]          [,81]
#> [1,] -5.884823e+191 1.977145e+194 -6.642683e+196 2.231766e+199 -7.498142e+201
#> [2,] -1.085147e+192 3.645808e+194 -1.224895e+197 4.115323e+199 -1.382639e+202
#>              [,82]          [,83]         [,84]          [,85]         [,86]
#> [1,] 2.519177e+204 -8.463769e+206 2.843602e+209 -9.553751e+211 3.209808e+214
#> [2,] 4.645303e+204 -1.560699e+207 5.243535e+209 -1.761689e+212 5.918809e+214
#>               [,87]         [,88]          [,89]         [,90]          [,91]
#> [1,] -1.078410e+217 3.623174e+219 -1.217290e+222 4.089774e+224 -1.374056e+227
#> [2,] -1.988563e+217 6.681046e+219 -2.244655e+222 7.541445e+224 -2.533726e+227
#>              [,92]          [,93]         [,94]          [,95]         [,96]
#> [1,] 4.616464e+229 -1.551010e+232 5.210982e+234 -1.750752e+237 5.882063e+239
#> [2,] 8.512649e+229 -2.860025e+232 9.608926e+234 -3.228345e+237 1.084638e+240
#>               [,97]         [,98]          [,99]        [,100]
#> [1,] -1.976218e+242 6.639568e+244 -2.230719e+247 7.494626e+249
#> [2,] -3.644098e+242 1.224321e+245 -4.113393e+247 1.381991e+250