Toeplitz Calculations

Multiplication of stacked data vector with a block Toeplitz matrix (toepl_fwd() for MA calculations) or "inversion" of a block Toeplitz matrix in order to perform calculations equivalent to multiplying a given stacked data vector with the inverse of a lower-triangular banded block Toeplitz matrix (toepl_inv() for AR calculations). Note that matrix polynomials can be mapped one-to-one to banded lower-triangular block Toeplitz matrices.

Usage

toepl_fwd(polm_wide, data_in, t0 = 1)

toepl_inv(polm_rev, data_in, t0 = 1)

Arguments

polm_wide

Wide matrix \(( d_0, d_1, \ldots, d_q )\) of dimension \((m \times (q+1) n)\) which represents a matrix polynomial \(d(z)\) of degree \(q\).

data_in

Data matrix of dimension \((dim(inputs) x nobs)\). Every column corresponds to one observation.

t0

Integer. Time index where calculations should start. Default set to 1. For AR calculations, \(degree + 1\) would be another smart option.

polm_rev

Wide matrix \((c_p, ... , c_1)\) of dimension \((n \times (q+1) n)\) with coefficients ordered in reverse direction, and zero-lag coefficient matrix not included. It represents a square polynomial matrix \(c(z)\) with \(c_0\) equal to the identity matrix and of degree \(p\).

Value

data_out Data matrix of dimension \((dim_out x n_obs)\)

MA-type Toeplitz calculations

Given a polynomial matrix of degree \(q\) and dimension \((m \times n)\), where \(m \geq n\), and given a "wide" input data matrix of dimension \((n \times nobs)\), where \(nobs\) is the number of observations such that each column corresponds to one observation and the number of columns is equal to the number of observations, we calculate a "wide" output data matrix of dimension \((m \times nobs)\).

The function name toepl_fwd stems from the multiplication of the "stacked" input data vector \((u_1', \ldots , u_{nobs}')'\) with a banded lower-triangular block Toeplitz matrix \(T\) of dimension \((nobs m \times nobs n)\) whose block elements depend only on the difference between the row- and column-index such that \(T_{i,j} = d_{i-j}\).

AR-type Toeplitz calcuations

Given a square polynomial matrix \(c(z)\) for which \(c_0\) is the identity matrix and given a wide data matrix y = data_in, obtain the solution u, a wide data matrix, of the Toeplitz equation $$T (y_1', \ldots , y_{nobs}')' = (u_1', ... , u_{nobs}')'$$

Note that the zero-lag coefficient is discarded and the coefficients are in reversed order since this simplifies computations and implementation.

Examples

p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_fwd(p, data)
#>           [,1]     [,2]     [,3]     [,4]      [,5]      [,6]      [,7]
#> [1,] -35.05413 122.4093 32.33197 23.40080 -140.9991 -278.3618 -200.0227
#> [2,] -76.43227 222.9411 50.03340 41.54931 -274.2953 -533.4930 -383.1974
#>           [,8]      [,9]     [,10]     [,11]      [,12]     [,13]     [,14]
#> [1,] -334.3543 -144.0283 -172.2246 -62.33118  -52.97848 -357.9832 -476.9522
#> [2,] -630.8153 -263.9452 -306.7330 -99.81826 -102.07409 -676.3232 -892.9667
#>           [,15]     [,16]     [,17]     [,18]    [,19]    [,20]      [,21]
#> [1,]  -550.2408 -487.1566 -261.5457 -31.29698 327.7533 160.8773  -74.33956
#> [2,] -1011.9046 -913.2394 -492.4754 -68.86695 617.8572 297.7309 -135.08121
#>           [,22]     [,23]     [,24]      [,25]     [,26]     [,27]     [,28]
#> [1,]  -538.2684 -362.3175 -246.6079  -69.32463 -167.2260 -284.1703  4.744184
#> [2,] -1006.8516 -669.6878 -458.3411 -140.75320 -319.3051 -538.0137 17.135358
#>          [,29]    [,30]     [,31]     [,32]     [,33]     [,34]     [,35]
#> [1,] -8.941897 278.0233 -1.321262  5.515009 -477.9331 -163.9486 -148.5204
#> [2,] -9.561806 535.7349  8.157037 41.975720 -865.6481 -282.8428 -267.5348
#>         [,36]     [,37]      [,38]     [,39]     [,40]     [,41]     [,42]
#> [1,]  64.9754 -180.3866  -95.38995  69.43865 -105.7511 -167.1772 -49.16714
#> [2,] 126.0136 -321.1460 -168.31603 138.82349 -205.6106 -313.0978 -88.62185
#>          [,43]     [,44]     [,45]     [,46]     [,47]    [,48]     [,49]
#> [1,]  74.49028 -189.2950 -366.3280 -453.4131 -360.1436 -227.108 -280.7616
#> [2,] 142.24321 -358.3924 -683.7868 -844.3438 -653.3827 -394.616 -498.7625
#>         [,50]      [,51]    [,52]     [,53]    [,54]      [,55]     [,56]
#> [1,] 143.3409  -79.67451 262.6210 -27.24914 25.75118  -71.03016 -185.8742
#> [2,] 290.1269 -128.60668 495.5401 -48.07460 35.82062 -130.82488 -341.1189
#>          [,57]    [,58]    [,59]    [,60]      [,61]     [,62]    [,63]
#> [1,] -139.0424 38.29507 225.8776 10.12065  -96.14332 -134.2067 324.5914
#> [2,] -242.0143 82.84753 423.7890 21.69106 -176.18104 -248.1971 600.6494
#>         [,64]    [,65]     [,66]     [,67]     [,68]     [,69]     [,70]
#> [1,] 423.6956 486.8932  60.83966  84.12854 -237.4002 -150.8155 -259.8206
#> [2,] 784.9108 905.2350 116.80695 170.81842 -444.0290 -281.0863 -489.5001
#>          [,71]      [,72]     [,73]     [,74]     [,75]     [,76]     [,77]
#> [1,] -130.4348  -89.34535 -199.0224 -449.5946 -403.4716 -162.5038  92.43052
#> [2,] -245.2253 -179.05775 -386.6744 -850.9120 -755.0332 -298.7031 165.56693
#>           [,78]     [,79]     [,80]    [,81]    [,82]    [,83]     [,84]
#> [1,]  -66.60848 -239.6173 -334.5101 40.13898 15.59849 53.65836 -270.5573
#> [2,] -145.02136 -468.5950 -639.4135 73.45729 26.11279 91.01191 -506.8290
#>          [,85]      [,86]     [,87]     [,88]   [,89]     [,90]    [,91]
#> [1,] -467.3606  -547.2691 -315.8910 -144.0754 1.16607  -49.5735 152.3350
#> [2,] -853.2599 -1010.3276 -574.6504 -270.8081 3.63942 -101.7371 276.5421
#>         [,92]    [,93]     [,94]     [,95]     [,96]    [,97]    [,98]
#> [1,] 290.6325 130.1955 -165.4963 -154.2781  63.37445 242.1052 18.35597
#> [2,] 539.5531 253.9921 -294.0088 -274.6314 124.42711 446.8419 18.31746
#>          [,99]    [,100]
#> [1,] -153.1036 -145.7916
#> [2,] -297.8760 -279.9255
p = test_polm(dim = c(2,2), degree = 2) %>% unclass()
dim(p) = c(2,2*3)
data = matrix(rnorm(2*100), 2, 100)
toepl_inv(p, data)
#>           [,1]      [,2]     [,3]      [,4]        [,5]          [,6]
#> [1,] 0.9016598 -159.4440 54002.21 -18141220  6094977884 -2.047751e+12
#> [2,] 0.4787514 -297.5516 99561.62 -33452043 11238993461 -3.776004e+12
#>              [,7]          [,8]         [,9]         [,10]        [,11]
#> [1,] 6.879902e+14 -2.311465e+17 7.765911e+19 -2.609141e+22 8.766022e+24
#> [2,] 1.268638e+15 -4.262286e+17 1.432015e+20 -4.811193e+22 1.616434e+25
#>              [,12]        [,13]         [,14]        [,15]         [,16]
#> [1,] -2.945151e+27 9.894929e+29 -3.324434e+32 1.116922e+35 -3.752562e+37
#> [2,] -5.430789e+27 1.824601e+30 -6.130178e+32 2.059577e+35 -6.919635e+37
#>             [,17]         [,18]        [,19]         [,20]        [,21]
#> [1,] 1.260762e+40 -4.235825e+42 1.423125e+45 -4.781324e+47 1.606398e+50
#> [2,] 2.324814e+40 -7.810761e+42 2.624209e+45 -8.816647e+47 2.962160e+50
#>              [,22]        [,23]         [,24]        [,25]         [,26]
#> [1,] -5.397073e+52 1.813274e+55 -6.092120e+57 2.046791e+60 -6.876677e+62
#> [2,] -9.952074e+52 3.343634e+55 -1.123372e+58 3.774234e+60 -1.268043e+63
#>             [,27]         [,28]        [,29]         [,30]        [,31]
#> [1,] 2.310381e+65 -7.762270e+67 2.607917e+70 -8.761912e+72 2.943771e+75
#> [2,] 4.260288e+65 -1.431344e+68 4.808937e+70 -1.615676e+73 5.428243e+75
#>              [,32]        [,33]         [,34]        [,35]         [,36]
#> [1,] -9.890290e+77 3.322876e+80 -1.116398e+83 3.750803e+85 -1.260170e+88
#> [2,] -1.823746e+78 6.127303e+80 -2.058612e+83 6.916391e+85 -2.323724e+88
#>             [,37]         [,38]        [,39]         [,40]         [,41]
#> [1,] 4.233839e+90 -1.422458e+93 4.779082e+95 -1.605645e+98 5.394543e+100
#> [2,] 7.807099e+90 -2.622979e+93 8.812514e+95 -2.960771e+98 9.947408e+100
#>               [,42]         [,43]          [,44]         [,45]          [,46]
#> [1,] -1.812424e+103 6.089264e+105 -2.045832e+108 6.873453e+110 -2.309298e+113
#> [2,] -3.342066e+103 1.122846e+106 -3.772465e+108 1.267448e+111 -4.258291e+113
#>              [,47]          [,48]         [,49]          [,50]         [,51]
#> [1,] 7.758631e+115 -2.606695e+118 8.757804e+120 -2.942390e+123 9.885653e+125
#> [2,] 1.430673e+116 -4.806683e+118 1.614918e+121 -5.425698e+123 1.822891e+126
#>               [,52]         [,53]          [,54]         [,55]          [,56]
#> [1,] -3.321318e+128 1.115875e+131 -3.749044e+133 1.259580e+136 -4.231854e+138
#> [2,] -6.124431e+128 2.057647e+131 -6.913148e+133 2.322635e+136 -7.803438e+138
#>              [,57]          [,58]         [,59]          [,60]         [,61]
#> [1,] 1.421791e+141 -4.776842e+143 1.604892e+146 -5.392014e+148 1.811574e+151
#> [2,] 2.621749e+141 -8.808382e+143 2.959383e+146 -9.942745e+148 3.340499e+151
#>               [,62]         [,63]          [,64]         [,65]          [,66]
#> [1,] -6.086409e+153 2.044872e+156 -6.870230e+158 2.308215e+161 -7.754993e+163
#> [2,] -1.122319e+154 3.770696e+156 -1.266854e+159 4.256294e+161 -1.430002e+164
#>              [,67]          [,68]         [,69]          [,70]         [,71]
#> [1,] 2.605472e+166 -8.753698e+168 2.941011e+171 -9.881018e+173 3.319761e+176
#> [2,] 4.804429e+166 -1.614161e+169 5.423154e+171 -1.822036e+174 6.121559e+176
#>               [,72]         [,73]          [,74]        [,75]          [,76]
#> [1,] -1.115352e+179 3.747287e+181 -1.258989e+184 4.22987e+186 -1.421124e+189
#> [2,] -2.056682e+179 6.909907e+181 -2.321546e+184 7.79978e+186 -2.620520e+189
#>              [,77]          [,78]         [,79]          [,80]         [,81]
#> [1,] 4.774602e+191 -1.604140e+194 5.389486e+196 -1.810725e+199 6.083555e+201
#> [2,] 8.804252e+191 -2.957996e+194 9.938083e+196 -3.338933e+199 1.121793e+202
#>               [,82]         [,83]          [,84]         [,85]          [,86]
#> [1,] -2.043914e+204 6.867009e+206 -2.307133e+209 7.751357e+211 -2.604251e+214
#> [2,] -3.768928e+204 1.266260e+207 -4.254299e+209 1.429332e+212 -4.802176e+214
#>              [,87]          [,88]         [,89]          [,90]         [,91]
#> [1,] 8.749594e+216 -2.939632e+219 9.876385e+221 -3.318204e+224 1.114829e+227
#> [2,] 1.613404e+217 -5.420611e+219 1.821182e+222 -6.118689e+224 2.055718e+227
#>               [,92]         [,93]          [,94]         [,95]          [,96]
#> [1,] -3.745530e+229 1.258399e+232 -4.227887e+234 1.420458e+237 -4.772364e+239
#> [2,] -6.906667e+229 2.320457e+232 -7.796123e+234 2.619291e+237 -8.800125e+239
#>              [,97]          [,98]         [,99]         [,100]
#> [1,] 1.603388e+242 -5.386959e+244 1.809876e+247 -6.080703e+249
#> [2,] 2.956609e+242 -9.933424e+244 3.337367e+247 -1.121267e+250